+++ NEWS.md
@@ -1,8 +1,6 @@
 # Changes in mirtCAT 0.6
 
-- the `raw_responses` vector returned for each participant is now a character vector instead of an integer,
-  and the previous category location form is now found in `raw_responses_location` (for use with functions 
-  such as `fscores()`)
+- the `raw_responses` vector returned for each participant is now a character vector instead of an integer
 
 - added `slider` Type input for `sliderInput` rating scale questions